+++ tests/BlockCypher/Test/Api/TXTest.php
@@ -338,6 +338,28 @@ class TXTest extends ResourceModelTestCase
      * @dataProvider mockProvider
      * @param TX $obj
      */
+    public function testDecode($obj, $mockApiContext)
+    {
+        $mockBlockCypherRestCall = $this->getMockBuilder('\BlockCypher\Transport\BlockCypherRestCall')
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $mockBlockCypherRestCall->expects($this->any())
+            ->method('execute')
+            ->will($this->returnValue(
+                TXTest::getJson()
+            ));
+
+        $result = $obj->decode("hexRawTx", array(), $mockApiContext, $mockBlockCypherRestCall);
+
+        $this->assertNotNull($result);
+        $this->assertEquals($result, TXTest::getObject());
+    }
+
+    /**
+     * @dataProvider mockProvider
+     * @param TX $obj
+     */
     public function testGetMultipleWithParams($obj, $mockApiContext)
     {
         $mockBlockCypherRestCall = $this->getMockBuilder('\BlockCypher\Transport\BlockCypherRestCall')
@@ -410,6 +432,6 @@ class TXTest extends ResourceModelTestCase
 
         $result = $obj->create($mockApiContext, $mockBlockCypherRestCall);
         $this->assertNotNull($result);
-        $this->assertInstanceOf($result, '\BlockCypher\Api\TXSkeleton');
+        $this->assertInstanceOf('\BlockCypher\Api\TXSkeleton', $result);
     }
 }